Rename s3 to storage
